// vendor/github.com/jaegertracing/jaeger/jaeger-ui/packages/plexus/src/Digraph/types.tsx
export type TLayerType = 'html' | 'svg';
export enum ELayerType {
  Html = 'html',
  Svg = 'svg',
}

export type TRendererUtils = {
  getGlobalId: (name: string) => string;
  getZoomTransform: () => ZoomTransform;
};
// ...
export type TSetOnContainer<T = any, U = any> = {
  setOnContainer?: TSetProps<TFromGraphStateFn<T, U>>;
};

type TKeyed = { key: string };

export type TDefEntry<T = Record<string, unknown>, U = Record<string, unknown>> = {
  renderEntry?: (
    graphState: TExposedGraphState<T, U>,
    entryProps: TAnyProps | null,
    id: string
  ) => React.ReactElement;
  localId: string;
  setOnEntry?: TSetProps<TFromGraphStateFn<T, U>>;
};

export type TRenderNodeFn<T = Record<string, unknown>> = (
  vertex: TLayoutVertex<T>,
  utils: TRendererUtils
) => React.ReactNode;

export type TRenderMeasurableNodeFn<T = Record<string, unknown>> = (
  vertex: TVertex<T>,
  utils: TRendererUtils,
  layoutVertex: TLayoutVertex<T> | null
) => React.ReactNode;

export type TMeasureNodeUtils = {
  layerType: 'html' | 'svg';
  getWrapperSize: () => { height: number; width: number };
  getWrapper: () => TOneOfTwo<{ htmlWrapper: HTMLDivElement | null }, { svgWrapper: SVGGElement | null }>;
};
// ...
type TNodesLayer<T = Record<string, unknown>, U = Record<string, unknown>> = TKeyed &
  TOneOfTwo<TNodeRenderer<T>, TMeasurableNodeRenderer<T>> &
  TSetOnContainer<T, U>;

type TStandaloneNodesLayer<T = Record<string, unknown>, U = Record<string, unknown>> = TNodesLayer<T, U> &
  (
    | { layerType: Extract<TLayerType, 'html'> }
    | {
        layerType: Extract<TLayerType, 'svg'>;
        defs?: TNonEmptyArray<TDefEntry<T, U>>;
      });

export type TEdgesLayer<T = Record<string, unknown>, U = Record<string, unknown>> = TKeyed &
  TSetOnContainer<T, U> & {
    edges: true;
    markerEndId?: string;
    markerStartId?: string;
    setOnEdge?: TSetProps<(edge: TLayoutEdge<U>, utils: TRendererUtils) => TAnyProps | null>;
  };

export type TStandaloneEdgesLayer<T = Record<string, unknown>, U = Record<string, unknown>> = TEdgesLayer<
  T,
  U
> & {
  defs?: TNonEmptyArray<TDefEntry<T, U>>;
  layerType: Extract<TLayerType, 'svg'>;
};

export type THtmlLayersGroup<T = Record<string, unknown>, U = Record<string, unknown>> = TKeyed &
  TSetOnContainer<T, U> & {
    layerType: Extract<TLayerType, 'html'>;
    layers: TNonEmptyArray<TNodesLayer<T, U>>;
  };

export type TSvgLayersGroup<T = Record<string, unknown>, U = Record<string, unknown>> = TKeyed &
  TSetOnContainer<T, U> & {
    layerType: Extract<TLayerType, 'svg'>;
    defs?: TNonEmptyArray<TDefEntry<T, U>>;
    layers: TNonEmptyArray<TOneOfTwo<TNodesLayer<T, U>, TEdgesLayer<T, U>>>;
  };

export type TLayer<T = Record<string, unknown>, U = Record<string, unknown>> = TOneOfFour<
  THtmlLayersGroup<T, U>,
  TSvgLayersGroup<T, U>,
  TStandaloneNodesLayer<T, U>,
  TStandaloneEdgesLayer<T, U>
>;